Added presentation, assignment 3, fixed a page
So guess who forgot to render the changes before uploading files to their website so the page doesn't show, this guy. so I hopefully fixed that as well as added new pages for the project proposal Presentation and added to the existing assignment 3 page. Might have to fix some things after class if they don't work right now.