Added project presentation and paper, move priject files, fixed assignment 5
Had some issues with reorganizing the project files and rendering as a result, but I final got it rendered so hopefully everything should be in order now. sorry for the delay with this update. also assignment 2 and the labs 1 and 2 now did not want to render because of a missing "rmarkdown" so I reinstall but still get "The rmarkdown package is not available in this R installation.
Install with install.packages("rmarkdown")" so ill just update for now and continue to work on it later.