Version: 0.0.1
I am basing this website on Quarto, and am using my previous website I made for my EPPS6323 as a template that I plan to alter to better fit my use case as my professional website. I have uploaded my files from the previous website as a starting point, plan to have to fix a number of things such as navigation and pages as I progress. Hopefully not too many things are completely broken when I import the previous site.